<ahem> Switch to php. Added codex links, having existential crisis
As I'm filling in this array of codex links, I'm wondering: what am I
doing here that's going to help someone? The great visualization
Michelle originally did is already here. Just pointing to the codex is
a start, but I think a good helper may be to add a loop example for
some items. I'm going to fill in these codex links for now and mull it
over.